+++ test/test_filter.py
@@ -1,8 +1,9 @@
 import asyncio
 from typing import Any, Generator
 
+import aioreactive as rx
 import pytest
-from aioreactive import AsyncAwaitableObserver, asyncrx, run
+from aioreactive import AsyncAwaitableObserver
 from aioreactive.testing import VirtualTimeEventLoop
 from expression.core import pipe
 
@@ -20,7 +21,7 @@ def event_loop() -> Generator[Any, Any, Any]:
 
 @pytest.mark.asyncio
 async def test_filter_happy() -> None:
-    xs = asyncrx.from_iterable([1, 2, 3])
+    xs = rx.from_iterable([1, 2, 3])
     result = []
 
     async def asend(value: int) -> None:
@@ -31,15 +32,15 @@ async def test_filter_happy() -> None:
         await asyncio.sleep(0.1)
         return value > 1
 
-    ys = pipe(xs, asyncrx.filter_async(predicate))
-    value = await run(ys, AsyncAwaitableObserver(asend))
+    ys = pipe(xs, rx.filter_async(predicate))
+    value = await rx.run(ys, AsyncAwaitableObserver(asend))
     assert value == 3
     assert result == [2, 3]
 
 
 @pytest.mark.asyncio
 async def test_filter_predicate_throws() -> None:
-    xs = asyncrx.from_iterable([1, 2, 3])
+    xs = rx.from_iterable([1, 2, 3])
     err = MyException("err")
     result = []
 
@@ -50,9 +51,9 @@ async def test_filter_predicate_throws() -> None:
         await asyncio.sleep(0.1)
         raise err
 
-    ys = pipe(xs, asyncrx.filter_async(predicate))
+    ys = pipe(xs, rx.filter_async(predicate))
 
     with pytest.raises(MyException):
-        await run(ys, AsyncAwaitableObserver(asend))
+        await rx.run(ys, AsyncAwaitableObserver(asend))
 
     assert result == []
